Module Provider

Parametrize your OCaml library with values that behave like objects but aren't.

A "provider" is a construct that implements a set of methods that an library typically needs in order to provide certain functionality to a client.

The module is divided into several submodules:

Class_id: To identify classes.

Class: For implementing classes.

Interface: Manages the set of classes that an object implements.

Private: Used for testing purposes.

This module is inspired by the Eio.Resource module and provides a way to parameterize code when a library either doesn't want to or can't commit to a specific implementation.

module Class_id : sig ... end

module Class : sig ... end

This module is used by providers to implement classes.

module Interface : sig ... end

Manipulating the set of classes implemented by a provider.

type -'tags t =

| T : {

t : 't;

interface : ('t, 'tags) Interface.t;

} -> 'tags t

A provider is a pair of a value and an interface for it. Think about t as the internal state of an object, and interface as the set of methods that the object implements.

module Private : sig ... end

This module is exported for testing purposes only.