Module Provider.Class

This module is used by providers to implement classes.

type _ t = private

| T : {

class_id : ('t, 'implementation, _) Class_id.t;

implementation : 'implementation;

} -> 't t

val implement : class_id:('t, 'implementation, _) Class_id.t -> 'implementation -> 't t

implement ~class_id (module Impl) returns a class that uses Impl as the implementation for class_id.

The tags associated with the class_id are ignored at this stage. The handling of the tags happens at the interface building stage, not at the granularity of each class. This means that the implement function focuses solely on creating the class, without considering the tags that indicate which classes are supported by the provider.

Dump & debug

val uid : _ t -> Class_id.Uid.t

val info : _ t -> Class_id.Info.t