Module Provider.Class_id

type ('t, 'implementation, 'tag) t = ..

Identifying a class within the provider system.

't is the internal state of the provider itself.

'implementation is the API that can be requested.

'tag is the tag (or tags) indicating the supported class. It's a phantom type designed to make Interface.lookup more type-safe.

The API requested is expected to be a module type (Eio supports single functions but this is discouraged through the use of this library).

Dump & debug

module Info : sig ... end

val info : (_, _, _) t -> Info.t

Indexation

module Uid : sig ... end

val uid : (_, _, _) t -> Uid.t

val same : (_, _, _) t -> (_, _, _) t -> Base.bool