Module Provider.Interface

Manipulating the set of classes implemented by a provider.

This module provides functions for creating an interface, as well as retrieving and extending the classes implemented by an interface, making it easy to manage the functionalities that a provider supports.

type ('t, -'tags) t

An interface is essentially a collection of classes that an object implements, each of which providing a specific set of functionalities (one class = one first-class module with type t = 't).

't is the internal state of the provider.

'tags indicate which functionality are supported by the provider. It is a phantom type using polymorphic variants. To give an example, in the tests for this library, we have two modules defining each their own tag:

module Directory_reader = struct
  type tag = [ `Directory_reader ]
end

module File_reader = struct
  type tag = [ `File_reader ]
end

Then, the type of the interface for a provider whose internal state is state, that would implement both functionalities would be:

(state, [ Directory_reader.tag | File_reader.tag ]) Provider.Interface.t

Building interfaces

val make : 't Class.t Base.list -> ('t, _) t

make classes create a new interface from a list of classes. It only keeps the last occurrence of each class. This means that the resulting interface will not contain any duplicate classes, and the order of the classes in the input list can affect its contents.

val classes : ('t, _) t -> 't Class.t Base.list

classes t returns a list of classes that the interface t implements. See also extend.

val extend : ('t, _) t -> with_:'t Class.t Base.list -> ('t, _) t

extend t ~with_ extends the interface t and returns a new interface that includes both the original and additional classes. The resulting interface only contains the last occurrence of each class identifier, prioritizing the rightmost elements in the combined list classes t @ with_.

Lookup

A lookup operation is used to retrieve the implementation of a specific class within an interface based on its class identifier.

val is_empty : ('t, _) t -> Base.bool

is_empty t checks if an interface t implements any classes. An empty interface may be created using make []. It will cause any lookup operation to fail. It can be useful for initializing data structures or providing a base case for algorithms that process interfaces.

val lookup : ('t, 'tags) t -> class_id:('t, 'implementation, 'tags) Class_id.t -> 'implementation

lookup t ~class_id retrieves the implementation for a given class from an interface.

If the provider has correctly exported their implementation using the appropriate tags, the compiler will ensure that this function does not fail in user code (a failure of this function would typically indicate a programming error in the provider's setup).

val lookup_opt : ('t, _) t -> class_id:('t, 'implementation, _) Class_id.t -> 'implementation Base.option

lookup_opt t ~class_id returns the implementation of the class (Some implementation) or indicates that the class is not implemented (None).

This is particularly useful in scenarios where a part of a program needs to adapt behavior at runtime based on whether certain functionalities are available or not.

val implements : ('t, _) t -> class_id:('t, _, _) Class_id.t -> Base.bool

implements t ~class_id says wether an interface implements a class. This is true iif lookup_opt t ~class_id returns Some _.