Module Provider.Private

This module is exported for testing purposes only.

Its interface may change in breaking ways without requiring a major version of the library to be minted. Use at your own risk.

module Interface : sig ... end