Module Private.Interface

val same_class_uids : ('t, _) Interface.t -> ('t, _) Interface.t -> Base.bool

same_class_uids i1 i2 checks if the class identifiers of two interfaces are the same and in the same order.

val cache : (_, _) Interface.t -> Class_id.Uid.t Base.option

Exported to test the caching strategy. Retains the most recently looked up class. Currently returns None for empty interface, and if the interface is not empty, returns the most recently looked up class (Some uid) or an arbitrary initial value.