Module Private.Interface

val same_trait_uids : ('t, _) Interface.t -> ('t, _) Interface.t -> Base.bool

same_trait_uids i1 i2 checks if the traits of two interfaces are the same and in the same order.

val cache : (_, _) Interface.t -> Trait.Uid.t Base.option

Exported to test the caching strategy. Retains the most recently looked up trait. Currently returns None for empty interface, and if the interface is not empty, returns the most recently looked up trait (Some uid) or an arbitrary initial value.